+++ custom/apps/cvsu/__init__.py
@@ -1,5 +1,3 @@
-from custom.apps.cvsu.new_reports import NewChildProtectionReport, NewCVSUPerformanceReport, \
-    NewCVSUPerformanceReportTrend, NewChildProtectionReportTrend
 from .reports import ChildProtectionReport, ChildProtectionReportTrend, CVSUPerformanceReport, \
     CVSUPerformanceReportTrend
 
@@ -11,13 +9,5 @@ CUSTOM_REPORTS = (
     ('Performance Evaluation', (
         CVSUPerformanceReport,
         CVSUPerformanceReportTrend,
-    )),
-    ('New Child Protection & Gender based Violence', (
-        NewChildProtectionReport,
-        NewChildProtectionReportTrend,
-    )),
-    ('New Performance Evaluation', (
-        NewCVSUPerformanceReport,
-        NewCVSUPerformanceReportTrend,
-    )),
+    ))
 )
