+++ custom/apps/cvsu/sqldata.py
@@ -1,31 +1,22 @@
 from operator import itemgetter
+
+from sqlagg.filters import BETWEEN
+
 from corehq.apps.groups.models import Group
 from corehq.apps.reports import util
 from corehq.apps.reports.standard import CommCareUserMemoizer
 from corehq.apps.reports.util import format_datatables_data, make_ctable_table_name
+from corehq.apps.userreports.sql import get_table_name
+from corehq.db import UCR_ENGINE_ID
+from custom.apps.cvsu.mixins import CVSUSqlDataMixin, FilterMixin, DateColumnMixin, combine_month_year, format_date, \
+    combine_quarter_year, format_year
+from custom.apps.cvsu.utils import dynamic_date_aggregation
 from .filters import ALL_CVSU_GROUP
 from dimagi.utils.decorators.memoized import memoized
 from sqlagg import AliasColumn
 from sqlagg.columns import SimpleColumn, YearColumn, MonthColumn, YearQuarterColumn, SumColumn
 from corehq.apps.reports.datatables import DataTablesColumnGroup
 from corehq.apps.reports.sqlreport import DatabaseColumn, AggregateColumn, SqlData
-
-
-def combine_month_year(year, month):
-    months = ['Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun', 'Jul', 'Aug', 'Sep', 'Oct', 'Nov', 'Dec']
-    return "%s %s" % (months[int(month) - 1], int(year)), int(year * 100 + month)
-
-
-def combine_quarter_year(year, quarter):
-    return "%s Q%s" % (int(year), int(quarter)), int(year * 10 + quarter)
-
-
-def format_year(year):
-    return format_datatables_data(int(year), int(year))
-
-
-def format_date(value):
-    return format_datatables_data(value[0], value[1])
 
 
 def make_trend(reportdata):
@@ -137,11 +128,7 @@ class BaseSqlData(SqlData):
 
     @property
     def filters(self):
-        filters = ['date between :startdate and :enddate']
-        if not self.group_by_district:
-            filters.append('"user_id" in :users')
-
-        return filters
+        return []
 
     @property
     def filter_values(self):
@@ -262,6 +249,18 @@ class ChildProtectionData(AgeGenderFilteredReport):
     table_name = make_ctable_table_name('cvsulive_UnicefMalawiFluff')
 
     @property
+    def filters(self):
+        return [BETWEEN('date_reported', 'startdate', 'enddate')] + super(ChildProtectionData, self).filters[1:]
+
+    @property
+    def engine_id(self):
+        return UCR_ENGINE_ID
+
+    @property
+    def table_name(self):
+        return get_table_name(self.config['domain'], 'unicef_malawi')
+
+    @property
     def columns(self):
         cat_group = DataTablesColumnGroup("Category of abuse")
         return [
@@ -272,7 +271,8 @@ class ChildProtectionData(AgeGenderFilteredReport):
             DatabaseColumn("Neglect", SumColumn('abuse_category_neglect_total'), header_group=cat_group),
             DatabaseColumn("Exploitation", SumColumn('abuse_category_exploitation_total'), header_group=cat_group),
             DatabaseColumn("Other", SumColumn('abuse_category_other_total'), header_group=cat_group),
-            DatabaseColumn("Total incidents reported", SumColumn('abuse_category_total_total'), header_group=cat_group)
+            DatabaseColumn("Total incidents reported", SumColumn('abuse_category_total_total'),
+                           header_group=cat_group)
         ]
 
 
@@ -284,27 +284,62 @@ class ChildrenInHouseholdData(AgeGenderFilteredReport):
     has_total_column = False
 
     @property
+    def filters(self):
+        filters = super(ChildrenInHouseholdData, self).filters[1:]
+        return [BETWEEN('date_reported', 'startdate', 'enddate')] + filters
+
+    @property
+    def engine_id(self):
+        return UCR_ENGINE_ID
+
+    @property
+    def table_name(self):
+        return get_table_name(self.config['domain'], 'unicef_malawi')
+
+    @property
     def columns(self):
         return [
             self.location_column,
-            DatabaseColumn("Children per household experiencing abuse", SumColumn('abuse_children_abused_total')),
-            DatabaseColumn("Total number of children in household", SumColumn('abuse_children_in_household_total')),
+            DatabaseColumn(
+                "Children per household experiencing abuse", SumColumn('abuse_children_abused_total')
+            ),
+            DatabaseColumn(
+                "Total number of children in household", SumColumn('abuse_children_in_household_total')
+            ),
         ]
 
 
-class CVSUActivityData(BaseSqlData):
+class CVSUActivityData(FilterMixin, BaseSqlData):
     title = 'Activities Performed'
     chart_x_label = 'CVSU Location'
     chart_y_label = 'Number of reports'
     table_name = make_ctable_table_name('cvsulive_UnicefMalawiFluff')
 
     @property
+    def engine_id(self):
+        return UCR_ENGINE_ID
+
+    @property
+    def table_name(self):
+        return get_table_name(self.config['domain'], 'unicef_malawi')
+
+    @property
     def columns(self):
         return [
             self.location_column,
-            DatabaseColumn("Incidents of Abuse", SumColumn('incidents_total')),
-            DatabaseColumn("Outreach activities", SumColumn('outreach_total')),
-            DatabaseColumn("IGA Reports", SumColumn('iga_total')),
+            DatabaseColumn(
+                "Incidents of Abuse",
+                SumColumn('incidents_total',
+                          filters=self.filters + [BETWEEN('date_reported', 'startdate', 'enddate')])
+            ),
+            DatabaseColumn(
+                "Outreach activities",
+                SumColumn('outreach_total', filters=self.filters + [BETWEEN('date', 'startdate', 'enddate')])
+            ),
+            DatabaseColumn(
+                "IGA Reports",
+                SumColumn('iga_total', filters=self.filters + [BETWEEN('start_date', 'startdate', 'enddate')])
+            ),
             AggregateColumn(
                 "Total", self.sum,
                 [AliasColumn('incidents_total'), AliasColumn('outreach_total')]),
@@ -314,53 +349,315 @@ class CVSUActivityData(BaseSqlData):
         return (no_incidents or 0) + (outreach or 0)
 
 
-class CVSUServicesData(BaseSqlData):
+class CVSUServicesData(FilterMixin, BaseSqlData):
     title = 'Services Provided'
     chart_x_label = 'CVSU Location'
     chart_y_label = 'Number of incidents'
     table_name = make_ctable_table_name('cvsulive_UnicefMalawiFluff')
 
     @property
+    def engine_id(self):
+        return UCR_ENGINE_ID
+
+    @property
+    def table_name(self):
+        return get_table_name(self.config['domain'], 'unicef_malawi')
+
+    @property
     def columns(self):
         return [
             self.location_column,
-            DatabaseColumn("Counselling", SumColumn('service_counselling_total')),
-            DatabaseColumn("Psychosocial Support", SumColumn('service_psychosocial_support_total')),
-            DatabaseColumn("First Aid", SumColumn('service_first_aid_total')),
-            DatabaseColumn("Shelter", SumColumn('service_shelter_total')),
-            DatabaseColumn("Referral", SumColumn('service_referral_total')),
-            DatabaseColumn("Mediation", SumColumn('service_mediation_total')),
-            DatabaseColumn("Other", SumColumn('service_other_total')),
-            DatabaseColumn("Total", SumColumn('service_total_total')),
+            DatabaseColumn(
+                "Counselling",
+                SumColumn('service_counselling_total',
+                          filters=self.filters + [BETWEEN('date_reported', 'startdate', 'enddate')])
+            ),
+            DatabaseColumn(
+                "Psychosocial Support",
+                SumColumn('service_psychosocial_support_total',
+                          filters=self.filters + [BETWEEN('date_reported', 'startdate', 'enddate')])
+            ),
+            DatabaseColumn(
+                "First Aid",
+                SumColumn('service_first_aid_total',
+                          filters=self.filters + [BETWEEN('date_reported', 'startdate', 'enddate')])
+            ),
+            DatabaseColumn(
+                "Shelter", SumColumn('service_shelter_total',
+                                     filters=self.filters + [BETWEEN('date_reported', 'startdate', 'enddate')])
+            ),
+            DatabaseColumn(
+                "Referral",
+                SumColumn('service_referral_total',
+                          filters=self.filters + [BETWEEN('date_reported_mediated', 'startdate', 'enddate')])
+            ),
+            DatabaseColumn(
+                "Mediation",
+                SumColumn('service_mediation_total',
+                          filters=self.filters + [BETWEEN('date_reported', 'startdate', 'enddate')])
+            ),
+            DatabaseColumn(
+                "Other",
+                SumColumn('service_other_total',
+                          filters=self.filters + [BETWEEN('date_reported', 'startdate', 'enddate')])
+            ),
+            DatabaseColumn(
+                "Total",
+                SumColumn('service_total_total',
+                          filters=self.filters + [BETWEEN('date_reported_mediated', 'startdate', 'enddate')])
+            ),
         ]
 
 
-class CVSUIncidentResolutionData(BaseSqlData):
+class CVSUIncidentResolutionData(FilterMixin, BaseSqlData):
     title = 'Incident Resolution'
     chart_x_label = 'CVSU Location'
     chart_y_label = 'Number of incidents'
     table_name = make_ctable_table_name('cvsulive_UnicefMalawiFluff')
 
     @property
+    def engine_id(self):
+        return UCR_ENGINE_ID
+
+    @property
+    def table_name(self):
+        return get_table_name(self.config['domain'], 'unicef_malawi')
+
+    @property
     def columns(self):
         return [
             self.location_column,
-            DatabaseColumn("Resolved at CVSU", SumColumn('resolution_resolved_at_cvsu_total')),
-            DatabaseColumn("Referred to TA", SumColumn('resolution_referred_ta_total')),
-            DatabaseColumn("Referred to TA Court", SumColumn('resolution_referral_ta_court_total')),
-            DatabaseColumn("Referred to Police", SumColumn('resolution_referral_police_total')),
-            DatabaseColumn("Referred to Social Welfare", SumColumn('resolution_referral_social_welfare_total')),
-            DatabaseColumn("Referred to NGO", SumColumn('resolution_referral_ngo_total')),
-            DatabaseColumn("Referred to Other", SumColumn('resolution_referral_other_total')),
-            DatabaseColumn("Unresolved", SumColumn('resolution_unresolved_total')),
-            DatabaseColumn("Case Withdrawn", SumColumn('resolution_case_withdrawn_total')),
-            DatabaseColumn("Other", SumColumn('resolution_other_total')),
-            DatabaseColumn("Total", SumColumn('resolution_total_total')),
+            DatabaseColumn(
+                "Resolved at CVSU",
+                SumColumn('resolution_resolved_at_cvsu_total',
+                          filters=self.filters + [BETWEEN('mediation_provided_date', 'startdate', 'enddate')])
+            ),
+            DatabaseColumn(
+                "Referred to TA",
+                SumColumn('resolution_referred_ta_total',
+                          filters=self.filters + [
+                              BETWEEN('date_reported_provided_mediated', 'startdate', 'enddate')])
+            ),
+            DatabaseColumn(
+                "Referred to TA Court",
+                SumColumn('resolution_referral_ta_court_total',
+                          filters=self.filters + [
+                              BETWEEN('date_reported_provided_mediated', 'startdate', 'enddate')])
+            ),
+            DatabaseColumn(
+                "Referred to Police",
+                SumColumn('resolution_referral_police_total',
+                          filters=self.filters + [
+                              BETWEEN('date_reported_provided_mediated', 'startdate', 'enddate')])
+            ),
+            DatabaseColumn(
+                "Referred to Social Welfare",
+                SumColumn('resolution_referral_social_welfare_total',
+                          filters=self.filters + [
+                              BETWEEN('date_reported_provided_mediated', 'startdate', 'enddate')])
+            ),
+            DatabaseColumn(
+                "Referred to NGO",
+                SumColumn('resolution_referral_ngo_total',
+                          filters=self.filters + [
+                              BETWEEN('date_reported_provided_mediated', 'startdate', 'enddate')])
+            ),
+            DatabaseColumn(
+                "Referred to Other",
+                SumColumn('resolution_referral_other_total',
+                          filters=self.filters + [
+                              BETWEEN('date_reported_provided_mediated', 'startdate', 'enddate')])
+            ),
+            DatabaseColumn(
+                "Unresolved",
+                SumColumn('resolution_unresolved_total',
+                          filters=self.filters + [BETWEEN('mediation_provided_date', 'startdate', 'enddate')])
+            ),
+            DatabaseColumn(
+                "Case Withdrawn",
+                SumColumn('resolution_case_withdrawn_total',
+                          filters=self.filters + [BETWEEN('mediation_provided_date', 'startdate', 'enddate')])
+            ),
+            DatabaseColumn(
+                "Other",
+                SumColumn('resolution_other_total',
+                          filters=self.filters + [
+                              BETWEEN('date_reported_provided_mediated', 'startdate', 'enddate')])
+            ),
+            DatabaseColumn(
+                "Total",
+                SumColumn('resolution_total_total',
+                          filters=self.filters + [
+                              BETWEEN('date_reported_provided_mediated', 'startdate', 'enddate')])
+            ),
         ]
 
 
-ChildProtectionDataTrend = make_trend(ChildProtectionData)
-ChildrenInHouseholdDataTrend = make_trend(ChildrenInHouseholdData)
-CVSUServicesDataTrend = make_trend(CVSUServicesData)
-CVSUActivityDataTrend = make_trend(CVSUActivityData)
-CVSUIncidentResolutionDataTrend = make_trend(CVSUIncidentResolutionData)
+class ChildProtectionDataTrend(DateColumnMixin, make_trend(ChildProtectionData)):
+
+    @property
+    def columns(self):
+        cols = super(ChildProtectionDataTrend, self).columns
+        cols[0] = self.date_column
+        return cols
+
+
+class ChildrenInHouseholdDataTrend(DateColumnMixin, make_trend(ChildrenInHouseholdData)):
+
+    @property
+    def columns(self):
+        cols = super(ChildrenInHouseholdDataTrend, self).columns
+        cols[0] = self.date_column
+        return cols
+
+
+class CVSUServicesDataTrend(DateColumnMixin, CVSUSqlDataMixin, make_trend(CVSUServicesData)):
+    @property
+    def columns(self):
+        cols = [
+            self.location_column,
+            dynamic_date_aggregation(DatabaseColumn(
+                "Counselling",
+                SumColumn('service_counselling_total',
+                          filters=self.filters + [BETWEEN('date_reported', 'startdate', 'enddate')])
+            ), date_column='date_reported'),
+            dynamic_date_aggregation(DatabaseColumn(
+                "Psychosocial Support",
+                SumColumn('service_psychosocial_support_total',
+                          filters=self.filters + [BETWEEN('date_reported', 'startdate', 'enddate')])
+            ), date_column='date_reported'),
+            dynamic_date_aggregation(DatabaseColumn(
+                "First Aid",
+                SumColumn('service_first_aid_total',
+                          filters=self.filters + [BETWEEN('date_reported', 'startdate', 'enddate')])
+            ), date_column='date_reported'),
+            dynamic_date_aggregation(DatabaseColumn(
+                "Shelter", SumColumn('service_shelter_total',
+                                     filters=self.filters + [BETWEEN('date_reported', 'startdate', 'enddate')])
+            ), date_column='date_reported'),
+            dynamic_date_aggregation(DatabaseColumn(
+                "Referral",
+                SumColumn('service_referral_total',
+                          filters=self.filters + [BETWEEN('date_reported_mediated', 'startdate', 'enddate')])
+            ), date_column='date_reported_mediated'),
+            dynamic_date_aggregation(DatabaseColumn(
+                "Mediation",
+                SumColumn('service_mediation_total',
+                          filters=self.filters + [BETWEEN('date_reported', 'startdate', 'enddate')])
+            ), date_column='date_reported'),
+            dynamic_date_aggregation(DatabaseColumn(
+                "Other",
+                SumColumn('service_other_total',
+                          filters=self.filters + [BETWEEN('date_reported', 'startdate', 'enddate')])
+            ), date_column='date_reported'),
+            dynamic_date_aggregation(DatabaseColumn(
+                "Total",
+                SumColumn('service_total_total',
+                          filters=self.filters + [BETWEEN('date_reported_mediated', 'startdate', 'enddate')])
+            ), date_column='date_reported_mediated'),
+        ]
+
+        cols[0] = self.date_column
+        return cols
+
+
+class CVSUActivityDataTrend(DateColumnMixin, CVSUSqlDataMixin, make_trend(CVSUActivityData)):
+    @property
+    def columns(self):
+        cols = [
+            self.location_column,
+            dynamic_date_aggregation(DatabaseColumn(
+                "Incidents of Abuse",
+                SumColumn('incidents_total', filters=self.filters + [
+                    BETWEEN('date_reported', 'startdate', 'enddate')] + self.filters),
+            ), date_column='date_reported'),
+            dynamic_date_aggregation(DatabaseColumn(
+                "Outreach activities",
+                SumColumn('outreach_total', filters=self.filters + [BETWEEN('date', 'startdate', 'enddate')])
+            ), date_column='date'),
+            dynamic_date_aggregation(DatabaseColumn(
+                "IGA Reports",
+                SumColumn('iga_total', filters=self.filters + [BETWEEN('start_date', 'startdate', 'enddate')])
+            ), date_column='start_date'),
+            AggregateColumn(
+                "Total", self.sum,
+                [AliasColumn('incidents_total'), AliasColumn('outreach_total')]),
+        ]
+
+        cols[0] = self.date_column
+        return cols
+
+
+class CVSUIncidentResolutionDataTrend(DateColumnMixin, CVSUSqlDataMixin, make_trend(CVSUIncidentResolutionData)):
+
+    @property
+    def columns(self):
+        cols = [
+            self.location_column,
+            dynamic_date_aggregation(DatabaseColumn(
+                "Resolved at CVSU",
+                SumColumn('resolution_resolved_at_cvsu_total',
+                          filters=self.filters + [BETWEEN('mediation_provided_date', 'startdate', 'enddate')])
+            ), date_column='mediation_provided_date'),
+            dynamic_date_aggregation(DatabaseColumn(
+                "Referred to TA",
+                SumColumn('resolution_referred_ta_total',
+                          filters=self.filters + [
+                              BETWEEN('date_reported_provided_mediated', 'startdate', 'enddate')])
+            ), date_column='date_reported_provided_mediated'),
+            dynamic_date_aggregation(DatabaseColumn(
+                "Referred to TA Court",
+                SumColumn('resolution_referral_ta_court_total',
+                          filters=self.filters + [
+                              BETWEEN('date_reported_provided_mediated', 'startdate', 'enddate')])
+            ), date_column='date_reported_provided_mediated'),
+            dynamic_date_aggregation(DatabaseColumn(
+                "Referred to Police",
+                SumColumn('resolution_referral_police_total',
+                          filters=self.filters + [
+                              BETWEEN('date_reported_provided_mediated', 'startdate', 'enddate')])
+            ), date_column='date_reported_provided_mediated'),
+            dynamic_date_aggregation(DatabaseColumn(
+                "Referred to Social Welfare",
+                SumColumn('resolution_referral_social_welfare_total',
+                          filters=self.filters + [
+                              BETWEEN('date_reported_provided_mediated', 'startdate', 'enddate')])
+            ), date_column='date_reported_provided_mediated'),
+            dynamic_date_aggregation(DatabaseColumn(
+                "Referred to NGO",
+                SumColumn('resolution_referral_ngo_total',
+                          filters=self.filters + [
+                              BETWEEN('date_reported_provided_mediated', 'startdate', 'enddate')])
+            ), date_column='date_reported_provided_mediated'),
+            dynamic_date_aggregation(DatabaseColumn(
+                "Referred to Other",
+                SumColumn('resolution_referral_other_total',
+                          filters=self.filters + [
+                              BETWEEN('date_reported_provided_mediated', 'startdate', 'enddate')])
+            ), date_column='date_reported_provided_mediated'),
+            dynamic_date_aggregation(DatabaseColumn(
+                "Unresolved",
+                SumColumn('resolution_unresolved_total',
+                          filters=self.filters + [BETWEEN('mediation_provided_date', 'startdate', 'enddate')])
+            ), date_column='mediation_provided_date'),
+            dynamic_date_aggregation(DatabaseColumn(
+                "Case Withdrawn",
+                SumColumn('resolution_case_withdrawn_total',
+                          filters=self.filters + [BETWEEN('mediation_provided_date', 'startdate', 'enddate')])
+            ), date_column='mediation_provided_date'),
+            dynamic_date_aggregation(DatabaseColumn(
+                "Other",
+                SumColumn('resolution_other_total',
+                          filters=self.filters + [
+                              BETWEEN('date_reported_provided_mediated', 'startdate', 'enddate')])
+            ), date_column='mediation_provided_date'),
+            dynamic_date_aggregation(DatabaseColumn(
+                "Total",
+                SumColumn('resolution_total_total',
+                          filters=self.filters + [
+                              BETWEEN('date_reported_provided_mediated', 'startdate', 'enddate')])
+            ), date_column='date_reported_provided_mediated'),
+        ]
+
+        cols[0] = self.date_column
+        return cols
